requirements.txt
streamlit==1.39.0
chromadb==0.5.5
sentence-transformers==3.1.1
google-generativeai==0.8.3
python-dotenv==1.0.1
numpy>=1.26.4
protobuf>=4.25.3
requests>=2.31.0
